run baisc asm

+++ src/lib/def.c
@@ -11,7 +11,7 @@ static bool def_unused_eq_fn(const def_data data_a, const def_data data_b) {
     return data_a.u64 == data_b.u64;
 }
 
-static void def_unused_print_fn(const def_data data, FILE *file, int32_t idnt, uint32_t print_opts) {
+static void def_unused_print_fn(const def_data data, FILE *file, int32_t idnt, uint64_t print_opts) {
     (void) data;
     (void) file;
     (void) idnt;
